# [HUDI-8301] Removed `DataItem` for Flink stream writing