### Change Logs

Currently, `StreamWriteFunction` uses buffer with `DataBucket`s consisting of `DataItem`s, which are transformed `HoodieRecord`s. And when we need to flush bucket, we convert `DataItem`'s back to `HoodieRecord`s in `DataBucket::writeBuffer`.

We could use `HoodieRecord`s directly to reduce number of objects in the heap, and decrease GC workload. In a result, Flink stream write is quicker in total.

|  | Before | After |
| -- | -- | -- |
| GC mainly related CPU samples | 66.7K | 43.6K |
| Memory allocation | 2.04 TB | 2.04 TB |
| Total time of 60M rows writing | 4 min | 3 min 50 s |

GC before:
![01 - GC before](https://github.com/user-attachments/assets/4a06fe98-eb20-4c51-b478-e3ff5cc5c99e)

GC after:
![02 - GC after](https://github.com/user-attachments/assets/4d9c4480-17f7-4cbe-b317-9bdf7d69f9c0)

CPU samples number, which are mainly related to GC, decreased from 66.7K samples to 43.6K:
![03 - GC compare](https://github.com/user-attachments/assets/5a73249b-06e6-4470-a0aa-924a867d9fc2)

Memory allocation in `StreamOneInputProcessor::processInput` also decreased by 0.2%:
![04 - memory allocation compare](https://github.com/user-attachments/assets/390019f5-3040-4cc7-84a1-86d79db4d8ce)

### Impact

Improved Flink stream write performance.

### Risk level (write none, low medium or high below)

Low

### Documentation Update

No need

### Contributor's checklist

- [x] Read through [contributor's guide](https://hudi.apache.org/contribute/how-to-contribute)
- [x] Change Logs and Impact were stated clearly
- [x] Adequate tests were added if applicable
- [x] CI passed