@geserdugarov 2024-10-04
Refactored setting of `FlinkOptions.WRITE_BATCH_SIZE` and `FlinkOptions.WRITE_TASK_MAX_SIZE` in unit tests.
Restarted CI by force push.